Refactor `SequenceFactory`, implement `Sequence` and `SequenceItem` classes;
- Move `get-sequence-files` under `sequence` subdir, from `helpers`
