+++ packages/11ty/_plugins/figures/helpers/index.js
@@ -1,5 +1,4 @@
 module.exports = {
-  getSequenceFiles: require('./get-sequence-files'),
   isCanvas: require('./is-canvas'),
   isImageService: require('./is-image-service'),
   isSequence: require('./is-sequence')
