+++ packages/11ty/_plugins/figures/sequence/get-sequence-files.js
@@ -1,6 +1,5 @@
 const fs = require('fs')
 const path = require('path')
-const isSequence = require('./is-sequence.js')
 
 /**
  * Figure `getSequenceFiles` helper
@@ -10,16 +9,14 @@ const isSequence = require('./is-sequence.js')
  * @param  {Object} iiifConfig IIIF Config data
  * @return {Array<string>}     An array of filenames
  */
-module.exports = (figure, iiifConfig) => {
-  if (!isSequence(figure)) return
-
-  const { sequences } = figure
+module.exports = (sequence, iiifConfig) => {
+  if (!sequence) return
   const { dirs } = iiifConfig
   const { imagesDir, inputRoot } = dirs
-  const sequenceDir = path.join(inputRoot, imagesDir, sequences[0].id)
+  const sequenceDir = path.join(inputRoot, imagesDir, sequence.id)
   const defaultSequenceRegex = /^\d{3}\.(jpg|png)$/
-  const sequenceRegex = sequences[0].regex
-    ? new RegExp(sequences[0].regex.slice(1, -1))
+  const sequenceRegex = sequence.regex
+    ? new RegExp(sequence.regex.slice(1, -1))
     : defaultSequenceRegex
   return fs
     .readdirSync(sequenceDir)
